Adds AnalysisRequests controller
- Moves creation logic into controller
- Adds route for displaying the details for a specific request (this is where we can display the results when they are ready along with the status of the request)
