+++ Freshli.Web/Controllers/HomeController.cs
@@ -42,18 +42,5 @@ namespace Freshli.Web.Controllers {
         }
       );
     }
-
-    [HttpPost]
-    public IActionResult Analysis(AnalysisRequest analysisRequest) {
-      var result = _db.AnalysisRequests.Add(analysisRequest);
-      _db.SaveChanges();
-
-      var runner = new AnalysisRunner(_db);
-      BackgroundJob.Enqueue(
-        () => runner.Run(analysisRequest.Id)
-      );
-
-      return View(result.Entity);
-    }
   }
 }
