+++ Freshli.Web/Startup.cs
@@ -76,6 +76,7 @@ namespace Freshli.Web {
 
       app.UseEndpoints(
         endpoints => {
+          endpoints.MapControllers();
           endpoints.MapControllerRoute(
             name: "default",
             pattern: "{controller=Home}/{action=Index}/{id?}"
