+++ Freshli.Web/Views/AnalysisRequests/Show.cshtml
@@ -1,4 +1,4 @@
-﻿@model AnalysisRequest
+@model AnalysisRequest
 
 <div>
     <p>GUID: @Model.Id</p>
